% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besceaApp.R
\name{besceaApp}
\alias{besceaApp}
\title{Bescea Shiny App}
\usage{
besceaApp(
  data,
  text_field,
  unique_id,
  modelname = NULL,
  searchname = "Bescea",
  results_count = 50,
  ...
)
}
\arguments{
\item{data}{Data frame, each document is a row/observation.}

\item{text_field}{Text field from data}

\item{unique_id}{Unique identifier from data}

\item{modelname}{Defaults to NULL. If you are loading data to a FastText model built using besceaBuildModel, put the name of model here.  The model should be saved as 3 different files in the 'models' folder.}

\item{searchname}{Optional name that appears at the top of the Shiny app}

\item{results_count}{Number of results to load to the Shiny app. This is also the number of results that will be exported to Excel if the data is downloaded. This number can be changed in the app as well.  You may be able to speed up performance by choosing a lower number.}
}
\description{
Instant search engine
}
\examples{
besceaApp(data = sneapsters[1:100,], 
          text_field = "post_text",
          unique_id = "textid",
          epochs = 1,
          min_word_count = 1,
          searchname = "test_search")
}
\keyword{text}
