% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besceaLoadData.R
\name{besceaLoadData}
\alias{besceaLoadData}
\title{Load new data to bescea model}
\usage{
besceaLoadData(
  data,
  text_field = text_field,
  unique_id = unique_id,
  modelname = NULL,
  searchname = "my_search",
  spacy_nlp_model = NULL,
  ...
)
}
\description{
Load new data to bescea model for searching
}
\examples{
# Build model called "my_model" first
besceaBuildModel(data = sneapsters[1:100,], 
  text_field = "post_text",
  unique_id = "textid", 
  min_word_count = 1,
  epochs = 1, 
  modelname = "my_model")
  
# Then load data and run using "my_model"  
besceaLoadData(data = sneapsters[1:100,], 
  text_field = "post_text",
  unique_id = "textid",
  modelname = "my_model")
}
\keyword{text}
