% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besceaSearch.R
\name{besceaSearch}
\alias{besceaSearch}
\title{Search Text}
\usage{
besceaSearch(search_query, results_count = 50, fasttext_weight = 0.5)
}
\arguments{
\item{search_query}{search query}

\item{results_count}{how many docs to return}
}
\description{
Function to search bescea model
}
\examples{
besceaSearch("hello")
}
\keyword{text}
