% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSneapLarge.R
\name{downloadSneaplarge}
\alias{downloadSneaplarge}
\title{Download Sneaplarge Data Set}
\usage{
downloadSneaplarge()
}
\description{
Load data from github. Installs into models folder.
}
\examples{
downloadSneapLarge()
}
\keyword{data}
\keyword{sneap}
\keyword{text}
