% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sneapsters.R
\docType{data}
\name{sneapsters}
\alias{sneapsters}
\title{Selection of Text Posts}
\format{
A data frame with 2000 rows and 2 variables:
\describe{
\item{textid}{unique id of each post}
\item{post_text}{text of the post. title has been removed.}
...
}
}
\source{
\url{https://www.ultimatemetal.com/forum/forums/andy-sneap/}
}
\usage{
sneapsters
}
\description{
A dataset containing posts from the Andy Sneap Forum.
}
\keyword{datasets}
