% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besceaApp.R
\name{besceaApp}
\alias{besceaApp}
\title{Bescea Shiny App}
\usage{
besceaApp(
  data,
  text_field,
  unique_id,
  modelname = NULL,
  searchname = "Bescea",
  results_count = 50,
  ...
)
}
\arguments{
\item{data}{the name of df}

\item{text_field}{x}

\item{unique_id}{x}

\item{modelname}{x}

\item{searchname}{x}

\item{results_count}{x}
}
\description{
App to search text
}
\examples{
besceaApp()
}
\keyword{text}
