% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besceaBuildModel.R
\name{besceaBuildModel}
\alias{besceaBuildModel}
\title{Build bescea model}
\usage{
besceaBuildModel(
  data,
  text_field,
  unique_id,
  modelname = "my_model",
  min_word_count = 1,
  epochs = 25,
  spacy_nlp_model = NULL,
  ...
)
}
\description{
Build bescea model
}
\examples{
besceaBuildModel(data = sneapsters[1:100,], 
  text_field = "post_text",
  unique_id = "textid", 
  min_word_count = 1,
  epochs = 1, 
  modelname = "my_model")
}
\keyword{text}
