% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besceaBuildModel.R
\name{besceaBuildModel}
\alias{besceaBuildModel}
\title{Build bescea model}
\usage{
besceaBuildModel(
  data,
  text_field,
  unique_id,
  modelname = "my_model",
  min_word_count = 1,
  epochs = 25,
  spacy_nlp_model = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame, each document is a row/observation.}

\item{text_field}{Text field from data, unquoted}

\item{unique_id}{Unique identifier from data, unquoted}

\item{modelname}{Your model name, to be referred to when loading new data. The model will be saved as 3 different files in the 'models' folder.  This value will be the \emph{modelname} arguement when running this model with \emph{besceaLoadData()} or \emph{besceaApp()}.}

\item{min_word_count}{Only consider tokens with at least n occurrences in the corpus}

\item{epochs}{Number of FastText epochs. More is generally better but takes longer.}

\item{spacy_nlp_model}{Defaults to \emph{NULL}.  When Python is run, SpaCy will load "en_core_web_sm" unless this argument is present. In that case, SpaCy will look for an nlp model in the location you provide.}

\item{...}{Arguments passed from other functions}
}
\description{
Build bescea model from a data frame containing a single text column and unique identifier column.
}
\examples{
besceaBuildModel(data = sneapsters[1:100,], 
  text_field = post_text,
  unique_id = textid, 
  min_word_count = 1,
  epochs = 1, 
  modelname = "my_model")
}
\keyword{engine}
\keyword{search}
\keyword{text}
