% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besceaLoadData.R
\name{besceaLoadData}
\alias{besceaLoadData}
\title{Load new data to bescea model}
\usage{
besceaLoadData(
  data,
  text_field = text_field,
  unique_id = unique_id,
  modelname = NULL,
  searchname = "my_search",
  spacy_nlp_model = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame, each document is a row/observation.}

\item{text_field}{Text field from data, unquoted}

\item{unique_id}{Unique identifier from data, unquoted}

\item{modelname}{Defaults to NULL. If you are loading data to a FastText model built using besceaBuildModel, put the name of model here.  The model should be saved as 3 different files in the 'models' folder.}

\item{searchname}{Optional name that appears at the top of the Shiny app}

\item{spacy_nlp_model}{Defaults to \emph{NULL}.  When Python is run, SpaCy will load "en_core_web_sm" unless this argument is present. In that case, SpaCy will look for an nlp model in the location you provide.}

\item{...}{Arguments passed from other functions}
}
\description{
Load new data to bescea model for searching
}
\examples{
# Build model called "my_model" first
besceaBuildModel(data = sneapsters[1:100,], 
  text_field = post_text,
  unique_id = textid, 
  min_word_count = 1,
  epochs = 1, 
  modelname = "my_model")
  
# Then load data and run using "my_model"  
besceaLoadData(data = sneapsters[1:100,], 
  text_field = post_text,
  unique_id = textid,
  modelname = "my_model")
}
\keyword{engine}
\keyword{search}
\keyword{text}
