% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besceaLoadData.R
\name{besceaLoadData}
\alias{besceaLoadData}
\title{Load new data to bescea model}
\usage{
besceaLoadData(
  data,
  text_field = text_field,
  unique_id = unique_id,
  modelname = NULL,
  searchname = "my_search",
  ...
)
}
\description{
Load new data to bescea model for seawrching
}
\examples{
besceaLoadData(...)
}
\keyword{text}
